feat(service): Initial implementation of server behavior added to project
Created a Point class representing a point on the coordinate plane , as well as a Cosine class that holds state information like Amplitude, Frequency, my Point class etc, and these are used to represent the Implanted Devices state that the client app will eventually change with gRPC calls.  Also created and implemented the server-side logic for requests and responses.